# docs/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Artifex API",
        "description": "二维游戏美术协作与 AI 资产生成平台 API",
        "version": "1.4.0",
        "contact": { "name": "Artifex Team" }
    },
    "servers": [
        { "url": "http://127.0.0.1:3000", "description": "本地开发" }
    ],
...
    "paths": {
...
        "/api/projects": {
            "get": {
                "tags": ["projects"],
                "summary": "获取用户所有项目",
                "parameters": [
                    { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1, "minimum": 1 }, "description": "页码" },
                    { "name": "pageSize", "in": "query", "schema": { "type": "integer", "default": 20, "minimum": 1, "maximum": 100 }, "description": "每页条数" }
                ],
                "responses": { "200": { "description": "项目列表" } }
            },
            "post": {
                "tags": ["projects"],
                "summary": "创建项目",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": ["name"],
                                "properties": {
                                    "name": { "type": "string", "description": "项目名称" },
                                    "description": { "type": "string", "description": "项目描述" },
                                    "type": { "type": "string", "description": "项目类型" }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": { "description": "创建成功" },
                    "400": { "description": "项目名称不能为空" }
                }
            }
        },
        "/api/projects/{id}": {
            "put": {
                "tags": ["projects"],
                "summary": "更新项目",
                "parameters": [
                    { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": { "type": "string" },
                                    "description": { "type": "string" },
                                    "type": { "type": "string" },
                                    "versionDesc": { "type": "string" }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": { "description": "更新成功" },
                    "404": { "description": "项目不存在" }
                }
            },
            "delete": {
                "tags": ["projects"],
                "summary": "删除项目",
                "parameters": [
                    { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
                ],
                "responses": { "200": { "description": "删除成功" } }
            }
        },
        "/api/projects/{id}/assets": {
            "post": {
                "tags": ["projects"],
                "summary": "添加项目素材",
                "description": "向项目添加素材。内容大小限制为 10MB。仅接受 PNG、JPG/JPEG、GIF、WebP、SVG 格式的图片。",
                "parameters": [
                    { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": ["name", "content"],
                                "properties": {
                                    "name": { "type": "string", "description": "素材名称" },
                                    "type": { "type": "string", "description": "素材类型（如 image/png, image/jpeg）" },
                                    "content": { "type": "string", "description": "素材内容（base64 或 data URL），最大 10MB" }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": { "description": "添加成功" },
                    "400": { "description": "素材名称和内容不能为空" },
                    "413": { "description": "素材内容过大，最大允许 10MB" }
                }
            }
        },
        "/api/assets/{id}": {
            "delete": {
                "tags": ["projects"],
                "summary": "删除项目素材",
                "parameters": [
                    { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
                ],
                "responses": { "200": { "description": "删除成功" } }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "cookieAuth": {
                "type": "apiKey",
                "in": "cookie",
                "name": "connect.sid"
            }
        }
    },
    "security": [{ "cookieAuth": [] }]
}